took TableJson out to table
